backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
smolagents>=1.0.0
litellm>=1.50.0
playwright>=1.49.0
google-cloud-firestore>=2.19.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
